+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Adds Storybook interaction tests for `RadioGroup`.
+
 ## 3.1.2 (May 9, 2024)
 
 ### Adds